+++ test/ex_zample_test.exs
@@ -181,4 +181,88 @@ defmodule ExZampleTest do
       assert user.email == "test@test.test"
     end
   end
+
+  describe "build_pair/1" do
+    import ExZample, only: [build_pair: 1]
+
+    test "uses struct default values" do
+      default_struct = %UserWithDefaults{}
+
+      assert {^default_struct, ^default_struct} = build_pair(UserWithDefaults)
+    end
+
+    test "uses the example values" do
+      default_struct = %UserWithDefaultsAndExample{}
+
+      assert {user_a, user_b} = build_pair(UserWithDefaultsAndExample)
+
+      assert default_struct != user_a
+      assert default_struct != user_b
+
+      assert user_a == %UserWithDefaultsAndExample{
+               first_name: "Example First Name",
+               last_name: "Example Last Name"
+             }
+
+      assert user_b == %UserWithDefaultsAndExample{
+               first_name: "Example First Name",
+               last_name: "Example Last Name"
+             }
+    end
+
+    test "works with structless modules" do
+      expected_example = %User{
+        age: 21,
+        email: "test@test.test",
+        first_name: "First Name",
+        id: 1,
+        last_name: "Last Name"
+      }
+
+      assert {^expected_example, ^expected_example} = build_pair(Factories.User)
+    end
+  end
+
+  describe "build_pair/2" do
+    import ExZample, only: [build_pair: 2]
+
+    test "overrides given attributes on struct default values" do
+      default_struct = %UserWithDefaults{}
+
+      {%UserWithDefaults{} = user_a, %UserWithDefaults{} = user_b} =
+        build_pair(UserWithDefaults, first_name: "Overrided First Name")
+
+      assert user_a.first_name == "Overrided First Name"
+      assert user_a.last_name == default_struct.last_name
+
+      assert user_b.first_name == "Overrided First Name"
+      assert user_b.last_name == default_struct.last_name
+    end
+
+    test "overrides given attributes on example values" do
+      {%UserWithDefaultsAndExample{} = user_a, %UserWithDefaultsAndExample{} = user_b} =
+        build_pair(UserWithDefaultsAndExample, first_name: "Overrided First Name")
+
+      assert user_a.first_name == "Overrided First Name"
+      assert user_a.last_name == "Example Last Name"
+
+      assert user_b.first_name == "Overrided First Name"
+      assert user_b.last_name == "Example Last Name"
+    end
+
+    test "overrides given attributes on structless modules" do
+      {%User{} = user_a, %User{} = user_b} =
+        build_pair(Factories.User, first_name: "Overrided First Name", age: 28)
+
+      assert user_a.first_name == "Overrided First Name"
+      assert user_a.age == 28
+      assert user_a.last_name == "Last Name"
+      assert user_a.email == "test@test.test"
+
+      assert user_b.first_name == "Overrided First Name"
+      assert user_b.age == 28
+      assert user_b.last_name == "Last Name"
+      assert user_b.email == "test@test.test"
+    end
+  end
 end